Added authentication for favourites

+++ lib/fizz_buzz_web/controllers/favourite_controller.ex
@@ -2,11 +2,10 @@ defmodule FizzBuzzWeb.FavouriteController do
   use FizzBuzzWeb, :controller
 
   alias FizzBuzz.Accounts
-  alias FizzBuzz.Accounts.Favourite
 
   def create(conn, %{"favourite" => favourite_params}) do
     case Accounts.create_favourite(favourite_params) do
-      {:ok, favourite} ->
+      {:ok, _favourite} ->
         conn
         |> put_flash(:info, "Favourite created successfully.")
         |> redirect(to: Routes.home_path(conn, :index))
@@ -20,10 +19,18 @@ defmodule FizzBuzzWeb.FavouriteController do
 
   def delete(conn, %{"id" => id}) do
     favourite = Accounts.get_favourite!(id)
-    {:ok, _favourite} = Accounts.delete_favourite(favourite)
 
-    conn
-    |> put_flash(:info, "Favourite deleted successfully.")
-    |> redirect(to: Routes.home_path(conn, :index))
+    if current_user_id(conn) == favourite.user_id do
+      {:ok, _favourite} = Accounts.delete_favourite(favourite)
+
+      conn
+      |> put_flash(:info, "Favourite deleted successfully.")
+      |> redirect(to: Routes.home_path(conn, :index))
+    else
+      conn
+      |> put_flash(:warning, "Page not found")
+      |> put_status(:not_found)
+      |> redirect(to: Routes.home_path(conn, :index))
+    end
   end
 end
